// src/users/dto/update-health.dto.ts
export class UpdateHealthInfoDto {
    height?: number;
    weight?: number;
    blood_type?: number;
    blood_pressure?: string;
    allergies?: string;
    chronic_conditions?: string;
    medical_history?: string;
    current_medications?: string;
    alcohol_consumption?: number;
    heart_rate?: number;
    body_temperature?: number;
} 